added pypi support and small renaming
