+++ mutate_nlp/__init__.py
@@ -1,18 +1,19 @@
-from mutate.pipelines.text_classification import TextClassificationSynthesize
+from typing import Dict, List, Optional, Union
+
 from transformers import PreTrainedModel
-from typing import List, Optional, Union, Dict
 
-SUPPORTED_TASKS = {
-    "text-classification-synthesis":TextClassificationSynthesize
-}
+from mutate_nlp.pipelines.text_classification import TextClassificationSynthesize
 
+SUPPORTED_TASKS = {"text-classification-synthesis": TextClassificationSynthesize}
 
 
-def pipeline(task:str,
-            model: Union[str, PreTrainedModel],
-            device: Optional[int] = -1,
-            generation_kwargs: Optional[Dict]={},
-            **kwargs):
+def pipeline(
+    task: str,
+    model: Union[str, PreTrainedModel],
+    device: Optional[int] = -1,
+    generation_kwargs: Optional[Dict] = {},
+    **kwargs,
+):
     """
 
     A factory method to return the pipeline corresponding to a given task.
@@ -26,12 +27,12 @@ def pipeline(task:str,
         task (str): Name of the task to synthesize or psuedo label examples. Currently supported tasks are:
 
             - `text-classification-synthesize` - will return `TextClassificationSynthesize` pipeline for synthesizing examples for text classification.
-        
+
         model (Union[str, PreTrainedModel]): Path or 🤗 model hub identifier. Currently supports Causal LM models only.
         device (Optional[int]): GPU Device to run the inference. Defaults to -1, runs on CPU.
-        generate_kwargs : Keyword arguments to override default generation params. This will be passed as generation params to model.generate - 
+        generate_kwargs : Keyword arguments to override default generation params. This will be passed as generation params to model.generate -
             https://github.com/huggingface/transformers/blob/fcb4f11c9232cee2adce8140a3a7689578ea97de/src/transformers/generation_utils.py#L803
-    
+
     Examples:
     --------
 
@@ -42,7 +43,7 @@ def pipeline(task:str,
                 model="EleutherAI/gpt-neo-2.7B",
                 device=1,
             )
-        
+
         Overriding the generation params:
 
         >> pipe = pipeline(
@@ -58,9 +59,9 @@ def pipeline(task:str,
                                     "early_stopping":True,
                                     "no_repeat_ngram_size":2
                 })
-        
+
         Using 🤗 Datasets:
-        
+
         >> task_desc = "Each item in the following contains customer service queries expressing the mentioned intent"
         >> synthesizerGen = pipe(
                         "banking77",
@@ -94,15 +95,13 @@ def pipeline(task:str,
     Returns:
         the corresponding pipeline class for the task
     """
-    
+
     if task in SUPPORTED_TASKS:
         pipeline_class = SUPPORTED_TASKS[task]
     else:
-        raise ValueError(f"Task - {task} is not supported. Supported tasks - {SUPPORTED_TASKS.keys()}")
+        raise ValueError(
+            f"Task - {task} is not supported. Supported tasks -"
+            f" {SUPPORTED_TASKS.keys()}"
+        )
 
-    return pipeline_class(model,device,**generation_kwargs)
-    
-
-    
-
-    
+    return pipeline_class(model, device, **generation_kwargs)
