+++ mutate_nlp/parsers/text_classification.py
@@ -1,9 +1,10 @@
-from mutate.parsers.parser import GenerationParser
-from typing import Dict, List, Optional
 import re
+from typing import Dict, List, Optional
+
+from mutate_nlp.parsers.parser import GenerationParser
+
 
 class TextClassificationSynthesizeParser(GenerationParser):
-
     regex_template = """.*{{label_title}}.*:.*{{label}}.*,.*{{example_title}}.*:(.*)"""
 
     def __init__(
@@ -16,7 +17,7 @@ class TextClassificationSynthesizeParser(GenerationParser):
         Parser to parse generated text to extract synthesized examples for a given class
 
         Args:
-            label_title (str): Name / title of the label that describes the classes. 
+            label_title (str): Name / title of the label that describes the classes.
                                Example - For sentiment classfication, the label title can be `Sentiment`
             example_title (str): Name / title of the label that describes the classes
                                  Example - For sentiment classfication, the example title can be `Review`
@@ -32,7 +33,7 @@ class TextClassificationSynthesizeParser(GenerationParser):
         """
         Args:
             generated_text (str): Generated text from the LLM
-            label (str): Class label 
+            label (str): Class label
 
         Returns:
             List[str]: List of parsed examples from the generated texts
@@ -42,9 +43,11 @@ class TextClassificationSynthesizeParser(GenerationParser):
         regex_pattern = self._get_regex(self.regex_template, self.template_params)
         return re.findall(regex_pattern, generated_text)
 
-class TextClassificationPseudolabelParser(GenerationParser):
 
-    regex_template = """.*{{label_title}}.*:.*(.*).*,.*{{example_title}}.*:.*{{example}}"""
+class TextClassificationPseudolabelParser(GenerationParser):
+    regex_template = (
+        """.*{{label_title}}.*:.*(.*).*,.*{{example_title}}.*:.*{{example}}"""
+    )
 
     def __init__(
         self,
@@ -53,11 +56,11 @@ class TextClassificationPseudolabelParser(GenerationParser):
         regex_template: Optional[str] = regex_template,
     ):
         """
-        Parser to parse generated text to extract lables generated for a given example 
+        Parser to parse generated text to extract lables generated for a given example
         by the text generation model
 
         Args:
-            label_title (str): Name / title of the label that describes the classes. 
+            label_title (str): Name / title of the label that describes the classes.
                                Example - For sentiment classfication, the label title can be `Sentiment`
             example_title (str): Name / title of the label that describes the classes
                                  Example - For sentiment classfication, the example title can be `Review`
